[Refactor] move type info to storage
