+++ be/src/storage/array_type_info.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "types/array_type_info.h"
+#include "storage/array_type_info.h"
 
 #include "base/utility/mem_util.hpp"
 #include "gutil/casts.h"
