+++ be/src/storage/array_type_info.h
