+++ be/src/storage/index/inverted/clucene/clucene_plugin.cpp
@@ -14,7 +14,7 @@
 
 #include "storage/index/inverted/clucene/clucene_plugin.h"
 
-#include "types/array_type_info.h"
+#include "storage/array_type_info.h"
 
 namespace starrocks {
