+++ be/src/storage/map_type_info.h
