+++ be/src/storage/rowset/default_value_column_iterator.cpp
@@ -44,11 +44,11 @@
 #include "column/column_builder.h"
 #include "column/datum.h"
 #include "runtime/decimalv3.h"
+#include "storage/array_type_info.h"
+#include "storage/map_type_info.h"
 #include "storage/range.h"
+#include "storage/struct_type_info.h"
 #include "storage/types.h"
-#include "types/array_type_info.h"
-#include "types/map_type_info.h"
-#include "types/struct_type_info.h"
 #include "util/json.h"
 #include "util/json_converter.h"
 #include "velocypack/Builder.h"
