+++ be/src/storage/struct_type_info.h
